might be relevant: more flexible input for calc_seasonality()
Constraints in input for function calc_seasonality() were a bit relaxed. Previously IDs had to be integer and column names of input matrices were predetermined which was not stated in the documentation. Now column names can be freely chosen (i.e. the order is relevant) and IDs can be of any type and will be substituted internally be integers to make it compliant with Fortran routine.