FIX: prof_class(): issues with specific user decisions
Re-arranged some parts of function prof_class() to make it work with specific user decisions (e.g. if only one TC per LU shall be produced which resulted in an error in the old version). In the meanwhile some small fixes have been made (typos in header of an output file and small addendum to manual).