ENH: area2catena(): improved efficiency
Improved efficiency (i.e. the function now runs much faster) by avoiding unnecessary calculations with the whole raster files which slowed function execution. Instead the extracted raster values are directly used which are much faster accessible.
NOTE: This may cause small deviations of results, i.e. slightly different mean catena lengths. This is as now for the calculation of the mean catena the distance to river values are used which have been corrected in case the smallest distance river value for an elementary hillslope is not zero but within 'max_riv_dist'.