FIX: useDynLib in NAMESPACE
In file NAMESPACE the compiled fortran code has to be specified by useDynLib(LUMP) which was commented out with a note to make it work under Windows. I don't know these reason for this anymore but in this way it did not work for me on Linux. However, it should also work like this under Windows...