FIX: calc_subbas()/lump_grass_post(): r.mapcalc formulas with R's paste()
For internally used GRASS function 'r.mapcalculator' sometimes the respective formulas are build with R's paste() function. This, however, converts large numbers into a scientific format, e.g. 100000 is pasted as 1e+05, which is not recognised by 'r.mapcalculator' and results in a runtime error. Therefore, this commit makes sure that large numbers are always pasted in decimal number format.