ENH: Added new function db_extract_subcatch()
This function extracts a sub-catchment based on a selected subbasin ID of an existing catchment using the processed WASA parameter database. It creates new databases for each desired sub-catchment.

Unfortunately the function is very slow in case of large basins, i.e. large amounts of datasets to process due to slow database operations with RODBC.