ENH: lump_grass_post(): Made calculation of subbasin statistics faster
ATTENTION: This comes at the cost of higher RAM usage (depending on the number of raster cells of the GRASS location). In the future, one could add the option to optimize for RAM usage (e.g. via writing/reading temporary raster data to/from disk) at the cost of higher computational burden.